+++ src/main/java/com/sparta/ticketauction/domain/user/response/SmsResponse.java
@@ -2,21 +2,17 @@ package com.sparta.ticketauction.domain.user.response;
 
 import java.time.LocalDateTime;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
+@RequiredArgsConstructor
 public class SmsResponse {
-	private String requestId;
-	private LocalDateTime requestTime;
-	private String statusCode;
-	private String statusName;
+	private final String requestId;
+	private final LocalDateTime requestTime;
+	private final String statusCode;
+	private final String statusName;
 	@Setter
 	private String verificationNumbers;
